moved LocationInput to resolvers/types.ts
