+++ src/graphql/resolvers/index.ts
@@ -2,3 +2,4 @@ export * from './book';
 export * from './location';
 export * from './person';
 export * from './resolvers';
+export * from './types';
