+++ src/graphql/resolvers/types.ts
@@ -1,0 +1,13 @@
+import { InputType, Field } from 'type-graphql';
+
+@InputType()
+export class LocationInput {
+  @Field({ nullable: true })
+  country: string;
+
+  @Field({ nullable: true })
+  city?: string;
+
+  @Field({ nullable: true })
+  place?: string;
+}
